Rename: profile api 구조 수정
