--- src/api/services/profile/mypage/index.ts
@@ -1,8 +1,0 @@
-import { fetchInstance } from '@/api/instance'
-import { MyPageItem } from '@/types'
-
-export const getMyPage = async (userId: string) => {
-  const response = await fetchInstance.get<MyPageItem>(`/api/profile/${userId}`)
-
-  return response.data
-}
